Moved click/json helpers to ext
