+++ appyter/helpers/fetch_and_serve.py
@@ -3,7 +3,8 @@ import sys
 import click
 from appyter import __version__
 from appyter.cli import cli
-from appyter.util import click_option_setenv, click_argument_setenv, join_routes
+from appyter.ext.click import click_option_setenv, click_argument_setenv
+from appyter.util import join_routes
 
 @cli.command(
   help='An alias for jupyter notebook which pre-fetches notebook data from a remote appyter and then launches serve',
