+++ appyter/profiles/default/fields/MultiChoiceField.py
@@ -1,5 +1,5 @@
 from appyter.fields import Field
-from appyter.util import try_json_loads
+from appyter.ext.json import try_json_loads
 
 class MultiChoiceField(Field):
   ''' Represing a multi-selectable combo box.
