+++ appyter/render/nbconstruct.py
@@ -10,8 +10,7 @@ from appyter.ext.fs import Filesystem
 from appyter.context import get_env, get_jinja2_env
 from appyter.parse.nb import nb_from_ipynb_io, nb_to_ipynb_io
 from appyter.parse.nbtemplate import cell_match
-from appyter.util import click_option_setenv, click_argument_setenv
-
+from appyter.ext.click import click_option_setenv, click_argument_setenv
 
 def render_cell(env, cell):
   ''' Render a single cell, calling jinja2 templates when necessary
