+++ appyter/util.py
@@ -6,18 +6,6 @@ from werkzeug.security import safe_join
 
 logger = logging.getLogger(__name__)
 
-def try_json_loads(v):
-  try:
-    return json.loads(v)
-  except:
-    return v
-
-def try_json_dumps(v):
-  if type(v) == str:
-    return v
-  else:
-    return json.dumps(v)
-
 def dict_filter_none(d):
   return { k: v for k, v in d.items() if v }
 
@@ -80,36 +68,6 @@ def importdir_deep(_dirname_, _package_, _globals_, filter_mod=lambda m, k, v: n
       for k, v in mod.__dict__.items()
       if filter_mod(mod, k, v)
     })
-
-def click_option_setenv(spec, envvar=None, **kwargs):
-  ''' Like click.option but explicitly set os.environ as well.
-  '''
-  import os, re, functools, click
-  m = re.match(r'^--(.+)$', spec)
-  assert m
-  var = m.group(1).replace('-', '_')
-  def decorator(func):
-    @click.option(spec, envvar=envvar, **kwargs)
-    @functools.wraps(func)
-    def wrapper(**kwargs):
-      if kwargs.get(var) is not None:
-        os.environ[envvar] = try_json_dumps(kwargs[var])
-      return func(**kwargs)
-    return wrapper
-  return decorator
-
-def click_argument_setenv(var, envvar=None, **kwargs):
-  ''' Like click.argument but explicitly set os.environ as well.
-  '''
-  import os, re, functools, click
-  def decorator(func):
-    @click.argument(var, envvar=envvar, **kwargs)
-    @functools.wraps(func)
-    def wrapper(**kwargs):
-      os.environ[envvar] = try_json_dumps(kwargs[var.replace('-', '_')])
-      return func(**kwargs)
-    return wrapper
-  return decorator
 
 def run_in_executor(f):
   import asyncio
